requirements.txt
# AILS Core Dependencies
# Created by Cherry Computer Ltd.

# ── Deep Learning ─────────────────────────────────────────────────────────────
tensorflow>=2.12.0
torch>=2.0.0

# ── Machine Learning ──────────────────────────────────────────────────────────
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# ── NLP ───────────────────────────────────────────────────────────────────────
nltk>=3.8.0
transformers>=4.30.0
tokenizers>=0.13.0

# ── Computer Vision ───────────────────────────────────────────────────────────
opencv-python-headless>=4.8.0
Pillow>=10.0.0

# ── Web Scraping ──────────────────────────────────────────────────────────────
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
lxml>=4.9.0
aiohttp>=3.8.0

# ── Database ──────────────────────────────────────────────────────────────────
mysql-connector-python>=8.0.33
pymongo>=4.4.0
sqlalchemy>=2.0.0

# ── Explainability (XAI) ──────────────────────────────────────────────────────
lime>=0.2.0.1
shap>=0.42.0

# ── Privacy ───────────────────────────────────────────────────────────────────
cryptography>=41.0.0

# ── API & Deployment ──────────────────────────────────────────────────────────
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
tqdm>=4.65.0
loguru>=0.7.0
PyYAML>=6.0.0
click>=8.1.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# ── Code Quality ──────────────────────────────────────────────────────────────
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
isort>=5.12.0
